requirements/test.txt
django-debug-toolbar==4.2.0
flake8==6.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-import-order==0.18.2
flake8-quotes==3.3.2
parameterized==0.9.0
pep8-naming==0.13.3